refactor(domain): replace old fare package with new fare2
- fix and update use cases
