+++ src/main/java/wooteco/subway/domain/path/fare/Decorator.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.path.fare2;
+package wooteco.subway.domain.path.fare;
 
 public abstract class Decorator implements Fare {
