--- src/main/java/wooteco/subway/domain/path/fare2/Fare.java
@@ -1,6 +1,0 @@
-package wooteco.subway.domain.path.fare2;
-
-public interface Fare {
-
-    int calculate();
-}
