+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,6 +3,9 @@ package wooteco.subway.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.path.fare.BasicFare;
+import wooteco.subway.domain.path.fare.DistanceOverFare;
+import wooteco.subway.domain.path.fare.Fare;
 import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.response.PathResponse;
@@ -24,8 +27,9 @@ public class PathService {
         Station startStation = stationRepository.findExistingStation(sourceStationId);
         Station endStation = stationRepository.findExistingStation(targetStationId);
         List<Section> sections = subwayRepository.findAllSections();
-        Path path = Path.of(startStation, endStation, sections);
 
-        return PathResponse.of(path);
+        Path path = Path.of(startStation, endStation, sections);
+        Fare fare = new DistanceOverFare(new BasicFare(), path.getDistance());
+        return PathResponse.of(path, fare.calculate());
     }
 }
